adds a jenky confirmation to deleting

+++ src/components/courses/CoursesPage.js
@@ -31,10 +31,13 @@ class CoursesPage extends React.Component {
   }
 
   handleDeleteCourse = (course) => {
-    toast.success("Course deleted");
-    this.props.actions.deleteCourse(course).catch((error) => {
-      toast.error("Delete failed. " + error.message, { autoClose: false });
-    });
+    const userWantsToDelete = confirm("Are you sure?");
+    if (userWantsToDelete) {
+      toast.success("Course deleted");
+      this.props.actions.deleteCourse(course).catch((error) => {
+        toast.error("Delete failed. " + error.message, { autoClose: false });
+      });
+    }
   };
 
   render() {
